+++ core/src/main/kotlin/com/gitlab/kordlib/core/exception/KordInitializationException.kt
@@ -1,0 +1,7 @@
+package com.gitlab.kordlib.core.exception
+
+class KordInitializationException : Exception {
+    constructor(message: String) : super(message)
+    constructor(cause: Throwable) : super(cause)
+    constructor(message: String, cause: Throwable) : super(message, cause)
+}